Add a "words" mode to the keywords page by prefix
"words" mode will group keywords by a prefix, which is then removed from the front of each tag:

e.g. keywords "names/name1", "names/name2", "names/name3" will be listed in a group called "names" as "name1", "name2", "name3".

The separator(s) for the prefix are in config, default is ',/':

$conf['tags_words_split_chars'] = '/,';